Category and Product

+++ src/main/java/com/said/apps/entities/Order.java
@@ -19,9 +19,10 @@ import jakarta.persistence.Table;
 public class Order implements Serializable {
 
 	private static final long serialVersionUID = 1L;
+	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 
 	@JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss'z'", timezone = "GMT")
 	private Instant moment;
@@ -35,7 +36,7 @@ public class Order implements Serializable {
 	public Order() {
 	}
 
-	public Order(Integer id, Instant moment, OrderStatus orderStatus, User client) {
+	public Order(Long id, Instant moment, OrderStatus orderStatus, User client) {
 		super();
 		this.id = id;
 		this.moment = moment;
@@ -43,11 +44,11 @@ public class Order implements Serializable {
 		this.client = client;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 
@@ -101,5 +102,5 @@ public class Order implements Serializable {
 			return false;
 		return true;
 	}
-
+	
 }
